etl/requirements.txt
pandas==2.2.3
requests
pyyaml
selenium==4.2.0
pymongo==4.10.1